# Optimize vector sorting with fx<, fx>