This commit creates fast paths for vector sorting when
the comparison predicate is either `fx<` or `fx>`.

In order to check if the predicate *really* is one of those,
we check if it is `eq` to the binding of the symbols `fx<` and
`fx>` in the immutable `SCHEME` module

Remark 1: Fast paths for `fx<=` and `fx>=` would be possible,
but I'm not sure many people would use those, and they'd
clutter the code.

Remark 2: Fast paths for `fl<` and `fl>` would certainly be
possible, but those are not in the immutable `SCHEME` module
(it's a SRFI), so things become more complicated.

The following is not really a benchmark, but hints at how large the speedup is.

```
stklos> 
(import (srfi 27))
(import (srfi 132))

(define v (make-vector 5000000))

(dotimes (i 5000000)
  (vector-set! v i (- 20000 (random-integer 50000))))

(define w (make-vector 5000000))

(dotimes (i 5000000)
  (vector-set! w i (- 20000 (random-integer 50000))))
;; v
;; w
stklos> (time (vector-stable-sort! fx< v))
Elapsed time: 900.284000000003 ms
stklos> (dotimes (i 5000000)
  (vector-set! v i (- 20000 (random-integer 50000))))
stklos> (time (vector-stable-sort! fx< v))
Elapsed time: 901.396999999997 ms
stklos> (time (vector-stable-sort! < w))
Elapsed time: 6557.967 ms
stklos> (dotimes (i 5000000)
  (vector-set! w i (- 20000 (random-integer 50000))))
stklos> (time (vector-stable-sort! < w))
Elapsed time: 6506.484 ms
```