@jpellegrini 2022-04-29
Oh - this also makes `Scheme_module` visible to all of STklos (it was `static` before).