@jpellegrini 2022-04-29
@egallesio -- this is a nice speedup, but I also thought it would perhaps be interesting, now that we have the immutable `SCHEME` module, trying this also:
- implement counting sort, for both `<`, `>` and `fx<`, `fx>`, in a similar way
- check at which vector size it becomes better than timsort, and use sounting sort when the predicate is one of those, *and* the vector is large enough.

What do you think?