@jpellegrini 2022-05-01
Probably `Scheme_module` should be renamed `STk_Scheme_module`? I can do that if you think so.