@egallesio 2022-05-02
Hi @jpellegrini,

I think you don't need to use the SCHEME module since you can access the function directly in C. For instance, `fx<` is implemented by the C function `STk_fxlt`. Furthermore, it permits to avoid the use of `find-module` and `symbol-value` and we don't need to rely on the presence of the SCHEME module.