@jpellegrini 2022-05-02
> I think you don't need to use the SCHEME

Ooooh, you're right! That was silly of me, the object already is a tagged procedure. Not even the extra argument is necessary. I'll change that.