@jpellegrini 2022-05-02
Wait, it's not yet working properly. I'll have to review this, it passes the tests, but crashes on some specific inputs on the command line.