@jpellegrini 2022-05-03
Hm, this doesn't seem to work:

```c
    if (STk_eq(less,STk_fxlt)==STk_true)
	fprintf(stderr,"fx<?\n");
    else
	fprintf(stderr,"NO fx<?\n");
```

It never seem to recognize `STk_fxlt`.

When `less` is `fx<`, this code:
```c
    fprintf(stderr,"less = %p %ul, STk_fxlt = %p %ul\n",
	    less,
	    (unsigned long) less,
	    STk_fxlt,
	    (unsigned long) STk_fxlt);
```

prints this:
```
less = 0x557dd328d140 3542667584l, STk_fxlt = 0x557dd31fd480 3542078592l
```