@jpellegrini 2022-05-03
Ok. Perhaps we really need the SCHEME module.

```c
    SCM fxl   = STk_symbol_value(STk_intern("fx<"), Scheme_module, NULL);

    fprintf(stderr,"STk_fxl = %p %lu\n",
	    (void*) fxl,
	    (unsigned long) fxl);

    fprintf(stderr,"less = %p %lu, STk_fxlt = %p %lu\n",
	    less,
	    (unsigned long) less,
	    STk_fxlt,
	    (unsigned long) STk_fxlt);
```
The result is:
```
STk_fxl = 0x55f3f821e140 94506328383808
less = 0x55f3f821e140 94506328383808, STk_fxlt = 0x55f3f818e480 94506327794816
```

Maybe it's OK to access the binding of `STk_fxlt` from the SCHEME module when SRFI-132 is initialized, and store it in a global variable, visible only in the SRFI? That would get everything working, I guess...