@egallesio 2022-05-04
Since you have two C pointers, you cannot do  `if (less == STk_fxlt)` ?