@jpellegrini 2022-05-04
> Since you have two C pointers, you cannot do `if (less == STk_fxlt)` ?

I tried, but when I pass `fx<` to the procedure, the pointers are different, for some reason:

```
    less = 0x55f3f821e140
STk_fxlt = 0x55f3f818e480
```

I don't know why this happens. It should be the same pointer, shouldn't it?