@egallesio 2022-05-04
> I don't know why this happens. It should be the same pointer, shouldn't it?

Yep! That is really weird. I'll look at it when merging the PR.