@egallesio 2022-05-19
Hi@jpellegrini,

I think that this PR is OK (I have the same ratio between fast path and normal path that you had in the first post).
If it is OK for you, we can close.