@jpellegrini 2022-05-19
> If it is OK for you, we can close.

Yes, I'm OK with it the way it is.

Maybe later it would be nice to also optimize:

* `fl<`, `fl>` when SRFI-144 is loaded
* `<` and `>` when the vector is a suitable `uvector`