@egallesio 2022-06-24
Hi @jpellegrini,

II had to change the call of `memcpy` in the  `gallop` function, since the tests didn't run on Alpine Linux (used when building the Docker image). With `memcpy`, the program crashed on an invalid instruction (probably a `SIGSEGV`, but I don't know how `musl` reports errors). 

I have not changed the other calls to `memcpy` in the rest of the file. I suppose that changing all the `memcpy` should not lower too much the performances.