@jpellegrini 2022-06-24
Hi @egallesio !
No problem, I think the difference between `memmove` and `memcpy` is not big.